# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: QWIZDS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /quizzes:
    get:
      operationId: listQuizzes
      tags: [quizzes]
      summary: List quizzes
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: visibility
          in: query
          schema:
            type: string
            enum: [public, private, all]
      responses:
        "200":
          description: List of quizzes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quiz"
    post:
      operationId: createQuiz
      tags: [quizzes]
      summary: Create a quiz
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuizBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quiz"
  /quizzes/{id}:
    get:
      operationId: getQuiz
      tags: [quizzes]
      summary: Get a quiz by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quiz
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizWithQuestions"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateQuiz
      tags: [quizzes]
      summary: Update a quiz
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateQuizBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quiz"
    delete:
      operationId: deleteQuiz
      tags: [quizzes]
      summary: Delete a quiz
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /quizzes/{id}/questions:
    get:
      operationId: listQuestions
      tags: [questions]
      summary: List questions for a quiz
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Question"
    post:
      operationId: addQuestion
      tags: [questions]
      summary: Add a question to a quiz
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuestionBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
  /quizzes/{id}/questions/{questionId}:
    patch:
      operationId: updateQuestion
      tags: [questions]
      summary: Update a question
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: questionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateQuestionBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
    delete:
      operationId: deleteQuestion
      tags: [questions]
      summary: Delete a question
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: questionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /discover:
    get:
      operationId: discoverQuizzes
      tags: [quizzes]
      summary: Discover public quizzes
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Public quizzes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quiz"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Quiz:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        visibility:
          type: string
          enum: [public, private]
        questionCount:
          type: integer
        timesPlayed:
          type: integer
        creatorId:
          type: integer
        creatorName:
          type: string
        coverImageUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - visibility
        - questionCount
        - timesPlayed
        - creatorId
        - creatorName
        - createdAt
        - updatedAt
    QuizWithQuestions:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        visibility:
          type: string
          enum: [public, private]
        questionCount:
          type: integer
        timesPlayed:
          type: integer
        creatorId:
          type: integer
        creatorName:
          type: string
        coverImageUrl:
          type: ["string", "null"]
        questions:
          type: array
          items:
            $ref: "#/components/schemas/Question"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - visibility
        - questionCount
        - timesPlayed
        - creatorId
        - creatorName
        - questions
        - createdAt
        - updatedAt
    CreateQuizBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        visibility:
          type: string
          enum: [public, private]
        coverImageUrl:
          type: string
      required:
        - title
        - visibility
    UpdateQuizBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        visibility:
          type: string
          enum: [public, private]
        coverImageUrl:
          type: string
    Question:
      type: object
      properties:
        id:
          type: integer
        quizId:
          type: integer
        questionText:
          type: string
        questionType:
          type: string
          enum: [multiple_choice, true_false, short_answer]
        options:
          type: array
          items:
            type: string
        correctAnswer:
          type: string
        explanation:
          type: ["string", "null"]
        timeLimit:
          type: integer
        points:
          type: integer
        imageUrl:
          type: ["string", "null"]
        orderIndex:
          type: integer
      required:
        - id
        - quizId
        - questionText
        - questionType
        - options
        - correctAnswer
        - timeLimit
        - points
        - orderIndex
    CreateQuestionBody:
      type: object
      properties:
        questionText:
          type: string
        questionType:
          type: string
          enum: [multiple_choice, true_false, short_answer]
        options:
          type: array
          items:
            type: string
        correctAnswer:
          type: string
        explanation:
          type: string
        timeLimit:
          type: integer
        points:
          type: integer
        imageUrl:
          type: string
        orderIndex:
          type: integer
      required:
        - questionText
        - questionType
        - options
        - correctAnswer
    UpdateQuestionBody:
      type: object
      properties:
        questionText:
          type: string
        questionType:
          type: string
          enum: [multiple_choice, true_false, short_answer]
        options:
          type: array
          items:
            type: string
        correctAnswer:
          type: string
        explanation:
          type: string
        timeLimit:
          type: integer
        points:
          type: integer
        imageUrl:
          type: string
        orderIndex:
          type: integer
...