# lib/api-zod/src/generated/types/createQuestionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * QWIZDS API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateQuestionBodyQuestionType } from "./createQuestionBodyQuestionType";

export interface CreateQuestionBody {
  questionText: string;
  questionType: CreateQuestionBodyQuestionType;
  options: string[];
  correctAnswer: string;
  explanation?: string;
  timeLimit?: number;
  points?: number;
  imageUrl?: string;
  orderIndex?: number;
}