# lib/api-zod/src/generated/types/quiz.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * QWIZDS API specification
 * OpenAPI spec version: 0.1.0
 */
import type { QuizVisibility } from "./quizVisibility";

export interface Quiz {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  category?: string | null;
  visibility: QuizVisibility;
  questionCount: number;
  timesPlayed: number;
  creatorId: number;
  creatorName: string;
  /** @nullable */
  coverImageUrl?: string | null;
  createdAt: Date;
  updatedAt: Date;
}